refactor(backends): extract slim-safe CLI backend
