+++ src/nexus/backends/connectors/gmail/transport.py
@@ -36,7 +36,7 @@ from typing import TYPE_CHECKING, Any
 
 import yaml
 
-from nexus.backends.connectors.cli.display_path import sanitize_filename
+from nexus.backends.base.cli_backend import sanitize_filename
 from nexus.backends.connectors.gmail.utils import (
     fetch_emails_batch,
     list_emails_by_folder,
